requirements.txt
# Multi-model pipeline
anthropic>=0.40.0
openai>=1.0.0
google-genai>=1.0.0

# Voice generation
elevenlabs>=1.0.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pyttsx3>=2.90
